+++ test/llm_engine.py
@@ -1,4 +1,3 @@
-import textwrap
 import openai
 import os
 
@@ -10,7 +9,6 @@ from langchain.embeddings import OpenAIEmbeddings
 from langchain.prompts import PromptTemplate
 from langchain.chains import RetrievalQA
 from CustomRetriever import CustomRetriever
-from write_to_md import write_test_to_md
 
 import config
 
@@ -27,7 +25,7 @@ class llm_engine:
     def setup(self):
         try:
             load_status = load_dotenv(find_dotenv())
-            print(f"load_dotenv: {load_status}/n")
+            print(f"load_dotenv: {load_status}\n")
         except Exception as e:
             print(e)
         openai.api_key = os.getenv("OPENAI_API_KEY")
@@ -79,48 +77,3 @@ class llm_engine:
     def ask_question(self, question):
         result = self.qa(question)
         return result
-
-
-# Function to get better output in terminal
-def wrap_text_preserve_newlines(text, width=110):
-    # Split the input text into lines based on newline characters
-    lines = text.split("\n")
-
-    # Wrap each line individually
-    wrapped_lines = [textwrap.fill(line, width=width) for line in lines]
-
-    # Join the wrapped lines back together using newline characters
-    wrapped_text = "\n".join(wrapped_lines)
-
-    return wrapped_text
-
-
-# prints response form LLM in terminal
-def process_llm_response(llm_response):
-    print(wrap_text_preserve_newlines(llm_response["result"]))
-    print("\n\nSources:")
-    # sort res by each elements metadata["score"] descending
-    llm_response["source_documents"] = sorted(
-        llm_response["source_documents"], key=lambda x: x.metadata["score"], reverse=True)
-    for source in llm_response["source_documents"]:
-        print(
-            f"{source.metadata['source'].split('/')[-1]} | score: {str(format(source.metadata['score'], '.4f'))} | har_mindretall: {str(source.metadata['har_mindretall'])}"
-        )
-    print()
-
-
-def main():
-    engine = llm_engine()
-    query_list = config.QUERYLIST
-
-    answer_list = []
-    for i, q in enumerate(query_list):
-        answer = engine.ask_question(q)
-        answer_list.append(answer)
-        print(f"Test: {i+1}")
-        process_llm_response(answer)
-    write_test_to_md(config.FILENAME, query_list, answer_list)
-
-
-if __name__ == "__main__":
-    main()
